+++ spec/lib/tmuxinator/cli_spec.rb
@@ -104,6 +104,17 @@ describe Tmuxinator::Cli do
             expect(instance).to receive(:start).with(*args)
             subject
           end
+
+          context "and the append option is passed" do
+            let(:args) { ["sample", "--append"] }
+
+            it "should call #start" do
+              instance = instance_double(cli)
+              expect(cli).to receive(:new).and_return(instance)
+              expect(instance).to receive(:start).with("sample", "--append")
+              subject
+            end
+          end
         end
 
         context "a thor command" do